--- commands/music/leave.js
@@ -1,34 +1,0 @@
-const { Command } = require('discord.js-commando');
-
-module.exports = class LeaveCommand extends Command {
-  constructor(client) {
-    super(client, {
-      name: 'leave',
-      aliases: ['end', 'stop'],
-      group: 'music',
-      memberName: 'leave',
-      guildOnly: true,
-      description: 'Leaves voice channel if in one!'
-    });
-  }
-
-  async run(message) {
-    var voiceChannel = message.member.voice.channel;
-    if (!voiceChannel)
-      return message.reply(
-        ':no_entry: Please join a voice channel and try again!'
-      );
-    if (!message.client.player.isPlaying(message)) {
-      return message.say(':x: There is no song playing right now!');
-    } else if (voiceChannel.id !== message.guild.me.voice.channel.id) {
-      message.reply(
-        `:no_entry: You must be in the same voice channel as the bot's in order to use that!`
-      );
-      return;
-    } else {
-      let song = message.client.player.stop(message);
-      if(song) 
-          message.channel.send(`:stop_button: Music stopped, and the Queue was cleared!`);
-     }
-   }
-}
