+++ commands/other/sus/messages/sus.message.js
@@ -3,5 +3,5 @@ const {baseEmbed} = LyraCore;
 module.exports = function () {
     return baseEmbed()
     .setTitle(`Kinda sus ngl...`)
-    .setImage('https://images2.imgbox.com/d3/27/BiAzzqzx_o.png')
+    .setImage('https://cdn.vox-cdn.com/thumbor/ot2xqOkZU-kuhd4G4fQWkjZs2CU=/0x60:1920x1020/fit-in/1200x600/cdn.vox-cdn.com/uploads/chorus_asset/file/21874606/Emergency_Meeting.jpg')
 }
